+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to SimpleBroker will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.0] - 2025-08-13
+### Changed
+- Added config keyword argument to functions that use config values
+- Allows easier testing, passing configs from API users
+- All existing code should work without change
 
 ## [2.4.0] - 2025-08-13
 ### Added